Create ingest_automation
use this script to ingest media from an artist-provided drive to a digital repository path and capture media info reports, md5s, and framemd5s. The md5s will be verified between the ingested media and digital repository. The metadata will be written to the artwork folder that is created and organized by this script